Generate RDS master password secret using a null resource and provisioner
Terraform's aws_secretsmanager_secret_version resource stores the secret value
in *plaintext* in the state file. We can workaround this weakness by using a
provisioner to generate this value using the AWS CLI. We do so on a `null_resource`
so that provisioning can be easily rerun and made dependent on the underlying secret
existing.